requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
awsebcli==3.18.0
Babel==2.8.0
bcrypt==3.1.7
blessed==1.17.4
botocore==1.15.42
cached-property==1.5.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9
distlib==0.3.1
Django==2.2.8
django-countries==6.1.2
django-cuser==2017.3.16
django-extra-fields==2.0.3
django-jet==1.0.8
django-modeltranslation==0.14.4
django-phonenumber-field==4.0.0
django-solo==1.1.3
django-tinymce==3.1.0
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
future==0.16.0
idna==2.7
importlib-metadata==1.6.0
isort==5.6.4
jet==1.0
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
modeltranslation==0.25
paramiko==2.7.1
pathspec==0.5.9
phonenumbers==8.12.5
pycparser==2.20
pylint==2.6.0
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.3.1
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
urllib3==1.24.3
virtualenv==20.0.35
wcwidth==0.1.9
websocket-client==0.57.0
wrapt==1.12.1
zipp==3.1.0
